refactor: move authentication logic to authController.js
