+++ routes/routes.js
@@ -2,17 +2,18 @@ const {Router} = require('express');
 const router = Router();
 
 const controller = require('../controllers/controller');
+const authController = require('../controllers/authcontroller');
 
 // GET home page
 router.get('/', controller.getHome);
 
-router.get("/login", controller.getLogin);
-router.post("/login", controller.postLogin);
+router.get("/login", authController.getLogin);
+router.post("/login", authController.postLogin);
 
-router.get("/signup", controller.getSignup);
-router.post("/signup", controller.handlePostSignup);
+router.get("/signup", authController.getSignup);
+router.post("/signup", authController.handlePostSignup);
 
-router.get("/logout", controller.handleLogOut);
+router.get("/logout", authController.handleLogOut);
 
 router.post("/messages/delete/:id", controller.handleDeleteMessage);
 router.post("/messages/new", controller.handleNewMessage);
